+++ apps/frontend/src/pages/project/project.tsx
@@ -119,14 +119,32 @@ const Project = (): JSX.Element => {
 	const [contributorToSplit, setContributorToSplit] =
 		useState<ContributorGetAllItemResponseDto | null>(null);
 
+	const hasViewPermission =
+		checkHasPermission(
+			[PermissionKey.VIEW_ALL_PROJECTS, PermissionKey.MANAGE_ALL_PROJECTS],
+			userPermissions,
+		) ||
+		checkIsProjectPermitted({
+			permissions: [
+				ProjectPermissionKey.VIEW_PROJECT,
+				ProjectPermissionKey.EDIT_PROJECT,
+				ProjectPermissionKey.MANAGE_PROJECT,
+			],
+			projectId,
+			projectUserPermissions,
+		});
+
 	useEffect(() => {
 		if (projectId) {
 			void dispatch(projectActions.getById({ id: projectId }));
-			void dispatch(
-				projectActions.loadAllContributorsByProjectId(Number(projectId)),
-			);
+
+			if (hasViewPermission) {
+				void dispatch(
+					projectActions.loadAllContributorsByProjectId(Number(projectId)),
+				);
+			}
 		}
-	}, [dispatch, projectId]);
+	}, [dispatch, hasViewPermission, projectId]);
 
 	useEffect(() => {
 		if (projectPatchStatus === DataStatus.FULFILLED) {
@@ -299,7 +317,7 @@ const Project = (): JSX.Element => {
 			combinedPermissions,
 		) &&
 		checkIsProjectPermitted({
-			permission: ProjectPermissionKey.MANAGE_PROJECT,
+			permissions: [ProjectPermissionKey.MANAGE_PROJECT],
 			projectId,
 			projectUserPermissions,
 		});
@@ -310,7 +328,7 @@ const Project = (): JSX.Element => {
 			combinedPermissions,
 		) &&
 		checkIsProjectPermitted({
-			permission: ProjectPermissionKey.EDIT_PROJECT,
+			permissions: [ProjectPermissionKey.EDIT_PROJECT],
 			projectId,
 			projectUserPermissions,
 		});
@@ -321,7 +339,9 @@ const Project = (): JSX.Element => {
 			combinedPermissions,
 		) || hasManagePermission;
 
-	const hasSetupAnalyticsPermission = hasManageAllProjectsPermission;
+	const hasSetupAnalyticsPermission =
+		hasManageAllProjectsPermission || hasEditPermission;
+
 	const hasEditContributorPermission = hasManageAllProjectsPermission;
 	const hasMergeContributorPermission = hasManageAllProjectsPermission;
 	const hasSplitContributorPermission = hasManageAllProjectsPermission;